Add search input to Search.jsx
Please take note that I will be copying a lot of this code from previously
written code but which did not use git. However I am trying to commit
this one piece of code at a time in a way that makes sense.

This adds the search input element with accompanying useState. Also
added a placeholder for where the createdTrees will appear, and a
button that will later link to a tree creation screen.