Add pathsArray and nodesArray to Edit
With temp arrays to allow rendering of Outline and KonvaCanvas for
testing. I also used the commenting convention / recipes given in
the "How To Code" course I took because it might help future readers.
I intend to do this for this whole application. I didn't work on web
pages on that course though so it feels a bit different using them for
web.