Add handleAdd logic
handleAdd opens an appropriate (ie node-modal or path-modal) modal
for the edit page. I decided that a modal would be a sensible way for
the user to add details to the node or path so that the tree and outline
being made would still be in view. This logic still needs to be tested
though, since I forgot to add my Modal component before this.