Add conditional rendering of Modal and comments
Clearly Modals should be rendered conditionally, in this case, in
response to a click that sets isModalVisible to true.
I added some comments to the Modal component that I merely copied from
a previous version I made. I had to stare at these parts for a bit to
understand them; comments should make it easier for my future self
to understand.