Add backgroundBox and styling to modal
The div with class backgroundBox serves to "deactivate" background
elements by "covering" them, rendering them unclickable. This is my
way of simulating the behavior I see in other modals.

Within Edit.jsx, I also placed the conditionally rendered modal
outside the edit-container; initially for styling purposes that
didn't make it to this commit, but it also just makes sense to
place it outside since it's supposed to be outside the normal
flow of the other components.