Add preliminary renderTree function and wishlist
Included with this commit are some object changes so that they work
with the function. I also tried implementing some of the simpler helpers.

This is a big function requiring many helpers that will take care
of rendering a tree object into a Babylon mesh. I tried dividing them
similarly to what was done in the How To Code - Simple Data course by
UBCx, to facilitate the coding process and improve readability, among
other reasons also described in that course.