Change Edit and package.json for D3Chart
Set aside the previous Edit.jsx file as an ignored "Edit-babylonver.jsx"
for a new Edit file that will make use of D3 instead of Babylon. I might
still make use of that code, that took me a lot of time to think about,
in the future. Package.json was changed accordingly to include d3 and
remove konva.