Remove ambiguity of the word "source"
fromNode needs to be "source" in order for D3Chart to recognize it
(along with toNode which needs to be target). The original code used
the word "source" for the element clicked for editing, so there was
ambiguity introduced. I removed it by using "clickedElement" for the
element clicked for editing. I also removed the need for the
"setSourceArray" method by replacing it with setTree, which sets the
tree, which includes whatever array we intended to set in the first
place.