Add isTreeRenderable(), helpers and test
I decided to use this function as a "guard" before the ForceGraph
rendering logic, so that the tree can still render automatically
but only when the current structure is valid. I feel like this would
be better UX-wise than the previous idea of having a "render tree"
button that the user will press every time they want the tree to render.