Change isTreeRenderable to include singular nodes
This allows D3Chart to render a display when there is a singular node
in the list. This seems appropriate since the D3 ForceGraph allows it.
The only reason isTreeRenderable exists is to prevent the errors that
arise when the D3 ForceGraph tries to work with incompatible trees (ie
trees that have disconnected nodes, or paths that don't have either a
source or a target).