Support submission of new trees
Submission of trees, whether a tree with that ID exists in the database
already ("old" tree) or not ("new" tree) will be done through the Edit
page. The interface for creating new trees and editing old ones
shouldn't be too different so I decided it should be done this way.

To enable submission of new trees, in addition to creating a submission
function, a useEffect was created that informs the code whether the
tree being edited already exists in the database or not. The isNewTree
flag is necessary to inform the code whether a POST request (for new
trees) or a PUT request (for old trees) should be used. The PUT request
will follow in a later commit.