Enable submission of new nodes
Create node CRUD methods at SkillTreesContext.jsx +
Create node "cleanup" and submission methods on Edit.jsx
= Enabling node submissions.

On "cleanup" methods: ie removing fx and fy. Any node can be a root
node for a themed tree, so these properties, which define the fixed x
and y coordinates of a node, are only added to a node while they're
being viewed as part of a themed skill tree such as in Edit.jsx (and
later in Tree.jsx), and thus not submitted to the database.