Enable opening of edit page from clicked node
This is my attempt at a more minimalist and intuitive design for the
home page. Initially I thought of rendering a "+" button beside the
clicked node, but that's because I forgot that I should display
information on the clicked node. I think both the information and the
"+" button should be around the same area as to not confuse the user.
Also, rendering the "+" beside a node might obscure nodes behind it.

This time around I worked on the styles with a mobile-first approach,
since I think that will be it's more frequent usage.