Add ModuleModal and AddLinkModal
ModuleModal is a form that enables the user to connect a learning
module to Skill Nodes selected at Edit.jsx.
AddLinkModal is a form that enables the user to add URLs to the text
that they're writing for the learning module. This is an important
component for adding Resource nodes to the Neo4j database, which is
in turn important for the goal of the app to increase the usage of
unique online resources represented by uniform resource locators (URLs).