Implement module submission
Module and target node submission happens at the same time since
the target nodes are entered in the module creation page anyways.
I feel like this would be easier and more intuitive for users too.
If they want to add optional descriptions to each skill node, they
may do so by clicking the node to edit it (to be implemented).