Enable submission at Edit.jsx
Redirects to Search.jsx to view the universalTree appended with user's
edits at Edit.jsx. Only the local appendedTree is appended (see changes
at SkillTreesContext) so only the user's changes will be seen. I
couldn't add the universalTree to the fetchUniversalTree dependency
array (results in infinite re-renders) so this was the solution I went
with. Upon reload however, fetchUniversalTree will grab the database
version to show any changes by other users over the internet, so this
probably won't be a problem.