Changes for deployment
Deploying to github pages for a free option. I tried deploying through
my repository's Settings page, but nothing would appear (not even an
error message on screen; it just says Error 404 on the console) even
after giving it a few hours. So I tried searching and followed the first
comment (by "Arnas") on this post:
https://dev.to/shashannkbawa/deploying-vite-app-to-github-pages-3ane