Create structure for Tree page components
Navigating to tree, the route includes a startNodeId and endNodeId so
that the tree could be easily shared. There is a TreePageChart, similar
to the other page-specific chart components that contain the page-
specific behavior for the chart. A TreeModuleView component displays
the module that the user will go over to take themselves from their
prerequisite skills to their objective skills, which is one of the
major features/use cases of this app.